[Chore] new file `userConfiguration.ts`
